Extract makeWindowFromFacet to avoid circular imports
