+++ test/compile/data/window.test.ts
@@ -2,11 +2,12 @@
 
 import {assert} from 'chai';
 import {WindowTransformNode} from '../../../src/compile/data/window';
+import {makeWindowFromFacet} from '../../../src/compile/data/windowFacet';
 import {Transform} from '../../../src/transform';
 
 describe('compile/data/window', () => {
   it('creates correct window nodes for calculating sort field of crossed facet', () => {
-    const window = WindowTransformNode.makeFromFacet(null, {
+    const window = makeWindowFromFacet(null, {
       row: {field: 'r', type: 'nominal'},
       column: {field: 'c', type: 'nominal', sort: {op: 'median', field: 'x'}}
     });
@@ -28,7 +29,7 @@ describe('compile/data/window', () => {
 
   it('does not create any window nodes for crossed facet', () => {
     assert.deepEqual(
-      WindowTransformNode.makeFromFacet(null, {
+      makeWindowFromFacet(null, {
         row: {field: 'a', type: 'nominal'}
       }),
       null
